Code refactor and test.
