+++ src/app/actions/auth/registerUser.js
@@ -1,10 +1,10 @@
 "use server";
 
-import dbConnect from "@/lib/dbConnect";
+import dbConnect, { collectionNames } from "@/lib/dbConnect";
 
 export const registerUser = async (payload) => {
    try{
-    const result = await dbConnect("test-user").insertOne(payload);
+    const result = await dbConnect(collectionNames.TEST_USER).insertOne(payload);
     return result;
    } catch (error) {
     console.log(error)
