+++ src/app/api/items/[id]/route.js
@@ -1,10 +1,10 @@
-import dbConnect from "@/lib/dbConnect"
+import dbConnect, { collectionNames } from "@/lib/dbConnect"
 import { ObjectId } from "mongodb";
 
 export async function GET( req, {params}) {
     
     const p = await params;
-    const singleData = await dbConnect("users").findOne({_id: new ObjectId(p.id)})
+    const singleData = await dbConnect(collectionNames.PRACTICE_DATA).findOne({_id: new ObjectId(p.id)})
    
     return Response.json(singleData)
   }
@@ -22,7 +22,7 @@ export async function GET( req, {params}) {
     const p = await params;
     const postedData = await req.json()
     const filter = {_id : new ObjectId(p.id)}
-    const updatedResponse = await dbConnect("users").updateOne(filter, {$set: {...postedData}}, {upsert: true})
+    const updatedResponse = await dbConnect(collectionNames.PRACTICE_DATA).updateOne(filter, {$set: {...postedData}}, {upsert: true})
    
     return Response.json(updatedResponse)
   }
