+++ src/app/api/items/route.js
@@ -1,9 +1,9 @@
-import dbConnect from "@/lib/dbConnect"
+import dbConnect, { collectionNames } from "@/lib/dbConnect"
 import { revalidatePath } from "next/cache"
 
 export async function GET() {
     
-    const data = await dbConnect("users").find({}).toArray()
+    const data = await dbConnect(collectionNames.TEST_USER).find({}).toArray()
    
     return Response.json(data)
   }
@@ -11,7 +11,7 @@ export async function GET() {
   export async function POST(req) {
     
     const postedData = await req.json()
-    const result = await dbConnect("users").insertOne(postedData)
+    const result = await dbConnect(collectionNames.PRACTICE_DATA).insertOne(postedData)
     revalidatePath("/products")
     return Response.json(result)
   }
