+++ src/app/page.js
@@ -5,9 +5,9 @@ import Hero from "@/components/Hero";
 import Templates from "@/components/Templates";
 import Testimonial from "@/components/Testimonial";
 import { getServerSession } from "next-auth";
-import { authOptions } from "./api/auth/[...nextauth]/route";
 import AiPowerAssistant from "@/components/AiPowerAssistant";
 import UserInfo from "@/components/UserInfo";
+import { authOptions } from "@/lib/authOptions";
 
 export default async function Home() {
   const session = await getServerSession(authOptions);
