+++ src/components/Navbar.js
@@ -19,7 +19,8 @@ const Navbar = () => {
     { name: "Templates", path: "/resumeTemplates" },
     { name: "About Us", path: "/aboutPage" },
     { name: "Contact Us", path: "/contactPage" },
-    { name: "Register", path: "/signupPage" },
+    // { name: "Register", path: "/signupPage" },
+    // { name: "Products", path: "/products" },
     // { name: "Add Products", path: "/products/add" }
   ];
