+++ src/lib/dbConnect.js
@@ -1,5 +1,10 @@
 import { MongoClient, ServerApiVersion } from 'mongodb';
 
+export const collectionNames = {
+    TEST_USER: "test-user",
+    PRACTICE_DATA: "users"
+}
+
 function dbConnect(collectionName) {
     const uri = process.env.MONGODB_URI
     // Create a MongoClient with a MongoClientOptions object to set the Stable API version
